feat: adopt feedback semantics for durable Loops (#290)
* feat: adopt loops paper feedback semantics

* fix: reconcile post-rebase runtime contracts

* fix: scope extension reads to workspaces (BUG-20260801-extension-cli-workspace-reads)

* docs: record loops paper QA verification

* docs: document loop config file fields

* fix: resolve loop review findings

* fix: review round

* docs: loops new design

+++ internal/daemon/daemon_review_and_fix_e2e_integration_test.go
@@ -346,7 +346,7 @@ func assertReviewAndFixRunContract(t testing.TB, detail contract.LoopRunResponse
 func requireReviewGenerationOutput(
 	t testing.TB,
 	detail contract.LoopRunResponse,
-	generation int,
+	generation int64,
 	nodeID string,
 ) contract.LoopGenerationOutput {
 	t.Helper()
